Inline client
