+++ rs/ic_os/vsock/src/client.rs
@@ -1,0 +1,86 @@
+use crate::protocol::{Command, Response};
+use mockall::automock;
+
+#[automock]
+pub trait VSockClient {
+    fn send_command(&self, command: Command) -> Response;
+}
+
+#[cfg(target_os = "linux")]
+pub use linux::*;
+
+#[cfg(target_os = "linux")]
+mod linux {
+    use super::*;
+    use crate::protocol::{Request, Response, parse_response};
+    use std::io::{Read, Write};
+    use vsock::{VMADDR_CID_HOST, VsockStream};
+
+    pub struct LinuxVSockClient {
+        port: u32,
+    }
+
+    impl LinuxVSockClient {
+        pub const DEFAULT_PORT: u32 = 19090;
+
+        pub fn with_port(port: u32) -> Self {
+            Self { port }
+        }
+    }
+
+    impl Default for LinuxVSockClient {
+        fn default() -> Self {
+            Self::with_port(Self::DEFAULT_PORT)
+        }
+    }
+
+    impl VSockClient for LinuxVSockClient {
+        fn send_command(&self, command: Command) -> Response {
+            let port = self.port;
+            let guest_cid = vsock::get_local_cid().map_err(|e| e.to_string())?;
+
+            let request = Request { guest_cid, command };
+
+            send_request_to_host_and_parse_response(&request, &port)
+        }
+    }
+
+    pub fn send_request_to_host_and_parse_response(request: &Request, port: &u32) -> Response {
+        let response_str = send_request_to_host(request, port)?;
+
+        parse_response(response_str.as_str())
+    }
+
+    pub fn send_request_to_host(request: &Request, port: &u32) -> Result<String, String> {
+        let mut stream = create_stream(port).map_err(|e| e.to_string())?;
+
+        let json_request = serde_json::to_string(request).map_err(|e| e.to_string())?;
+        stream
+            .write_all(json_request.as_bytes())
+            .map_err(|e| e.to_string())?;
+
+        let read_result = read_response_from_host(&mut stream);
+
+        read_result.map_err(|e| e.to_string())
+    }
+
+    fn read_response_from_host(stream: &mut VsockStream) -> anyhow::Result<String> {
+        // 64 KiB - generous for current responses (typically <1 KiB) while
+        // preventing unbounded allocation from a misbehaving host.
+        const MAX_RESPONSE_SIZE: u64 = 64 * 1024;
+        let mut response = String::new();
+        stream
+            .take(MAX_RESPONSE_SIZE)
+            .read_to_string(&mut response)?;
+        Ok(response)
+    }
+
+    fn create_stream(port: &u32) -> Result<VsockStream, std::io::Error> {
+        let stream = VsockStream::connect_with_cid_port(VMADDR_CID_HOST, *port)?;
+        // Set a long timeout, so HostOS has enough time to upgrade.
+        stream.set_write_timeout(Some(std::time::Duration::from_secs(60 * 5)))?;
+        stream.set_read_timeout(Some(std::time::Duration::from_secs(60 * 5)))?;
+
+        Ok(stream)
+    }
+}
